Uses `tokio` for `event_loop` example.
