+++ Cargo.toml
@@ -11,20 +11,9 @@ test = false
 [dependencies]
 nginee_event_loop = { version = "0.1.0", path = "crate/event_loop", optional = true }
 
-# `smol` still needs work for WASM support.
-# https://github.com/stjepang/smol/issues/87
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen = "0.2.62"
-
-[dev-dependencies]
-smol = "0.1.4"
-
 [workspace]
-members = ["crate/*"]
+members = ["crate/*", "examples/*"]
 
 [features]
 default = ["event_loop"]
 event_loop = ["nginee_event_loop"]
-
-[patch.crates-io]
-smol = { git = "https://github.com/jjl/smol.git", branch = "feature/wasm" }
