+++ examples/event_loop/index.html
@@ -1,0 +1,15 @@
+<html>
+  <head>
+    <meta content="text/html;charset=utf-8" http-equiv="Content-Type"/>
+  </head>
+  <body>
+    <script type="module">
+      import init, * as exports from './pkg/example_event_loop.js';
+
+      window.onload = async function() {
+        await init();
+        exports.run_ignore_errors();
+      };
+    </script>
+  </body>
+</html>
