+++ examples/event_loop/src/bin/main.rs
@@ -1,0 +1,5 @@
+use example_event_loop::{run, Error};
+
+fn main() -> Result<(), Error> {
+    run()
+}
