+++ examples/event_loop/src/lib.rs
@@ -1,6 +1,9 @@
 use nginee::event_loop::{EventHandler, EventHandlingOutcome, EventLoop};
+use tokio::runtime::Runtime;
 
-use crate::error::Error;
+pub use crate::error::Error;
+
+mod error;
 
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
@@ -17,7 +20,7 @@ fn log(s: &str) {
     println!("{}", s);
 }
 
-fn countdown(mut count: u32) -> EventHandler<Error> {
+pub fn countdown(mut count: u32) -> EventHandler<Error> {
     EventHandler::<Error>::new(move || {
         count -= 1;
         async move {
@@ -32,23 +35,16 @@ fn countdown(mut count: u32) -> EventHandler<Error> {
     })
 }
 
-fn main() -> Result<(), Error> {
+/// Runs the application.
+pub fn run() -> Result<(), Error> {
     let event_loop = EventLoop::new(vec![countdown(10)]);
 
-    smol::run(event_loop.run())
+    let mut rt = Runtime::new()?;
+
+    rt.block_on(event_loop.run())
 }
 
-mod error {
-    use std::fmt;
-
-    #[derive(Debug)]
-    pub struct Error;
-
-    impl fmt::Display for Error {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, "Error")
-        }
-    }
-
-    impl std::error::Error for Error {}
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
+pub fn run_ignore_errors() {
+    let _ = run();
 }
